+++ src/webR/webr-main.ts
@@ -25,6 +25,7 @@ import {
   EvalROptions,
   FSMessage,
   FSMountMessage,
+  FSSyncfsMessage,
   FSReadFileMessage,
   FSWriteFileMessage,
   InstallPackagesOptions,
@@ -97,7 +98,7 @@ export type FSNode = {
 };
 
 /** An Emscripten Filesystem type */
-export type FSType = 'NODEFS' | 'WORKERFS';
+export type FSType = 'NODEFS' | 'WORKERFS' | 'IDBFS';
 
 /**
  * Configuration settings to be used when mounting Filesystem objects with
@@ -472,6 +473,10 @@ export class WebR {
       mountpoint: string
     ): Promise<void> => {
       const msg: FSMountMessage = { type: 'mount', data: { type, options, mountpoint } };
+      await this.#chan.request(msg);
+    },
+    syncfs: async (populate: boolean): Promise<void> => {
+      const msg: FSSyncfsMessage = { type: 'syncfs', data: { populate } };
       await this.#chan.request(msg);
     },
     readFile: async (path: string, flags?: string): Promise<Uint8Array> => {